Added Empires.
Galaxies have lists of them, systems are owned by them. They have their
own colors. Colors and owned systems are random for now. Name is
displayed in brackets in their color in the main UI. Also added brackets
([]) to the font. To differentiate them with the C character, the front
bar in [, and the back bar in ], are double width.